Abstract parse option builder
* and add it to import configs, and sources
  which both have parse options
